Migrate map logic to useMap hook
